修改地磅的查询语句.

+++ uip-web/src/main/java/com/sq/loadometer/service/TradeDataService.java
@@ -3,6 +3,7 @@ package com.sq.loadometer.service;
 import com.sq.entity.search.MatchType;
 import com.sq.entity.search.Searchable;
 import com.sq.inject.annotation.BaseComponent;
+import com.sq.loadometer.component.DblinkConnecter;
 import com.sq.loadometer.component.JdbcHelper;
 import com.sq.loadometer.domain.LoadometerIndicatorDto;
 import com.sq.loadometer.domain.Trade;
@@ -12,7 +13,6 @@ import com.sq.quota.domain.QuotaInstance;
 import com.sq.quota.domain.QuotaTemp;
 import com.sq.quota.repository.QuotaInstanceRepository;
 import com.sq.quota.repository.QuotaTempRepository;
-import com.sq.quota.service.QuotaComputInsService;
 import com.sq.service.BaseService;
 import com.sq.util.DateUtil;
 import org.slf4j.Logger;
@@ -25,7 +25,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.sql.SQLException;
 import java.text.ParseException;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.HashMap;
 import java.util.List;
 
@@ -56,9 +55,6 @@ public class TradeDataService extends BaseService<Trade, Long> {
     @Autowired
     private QuotaInstanceRepository quotaInstanceRepository;
 
-    @Autowired
-    private QuotaComputInsService quotaComputInsService;
-
     /**
      * 地磅流水数据同步
      */
@@ -87,12 +83,35 @@ public class TradeDataService extends BaseService<Trade, Long> {
 
         StringBuilder insertTradeBuilder = new StringBuilder();
         insertTradeBuilder
-                .append(" select * from Trade where productNet is not null and datastatus = 1 and CONVERT(varchar(12) , seconddatetime, 112 ) = ")
+                .append(" select ")
+                .append("       t.lsh AS lsh,  ")
+                .append("       t.ch AS carNo, ")
+                .append("       t.hm AS proCode, ")
+                .append("       t.dwdw AS sourceArea, ")
+                .append("       t.cmrs AS firstWeightTime, ")
+                .append("       t.cprs AS secondWeightTime, ")
+                .append("       t.mz AS gross, ")
+                .append("       t.pz AS tare, ")
+                .append("       t.jz AS net, ")
+                .append("       t.czy AS operator, ")
+                .append("       CONVERT (VARCHAR(12), t.cprs, 112) AS statDateNum ")
+                .append("    FROM   ")
+                .append("       czb t ")
+                .append("    WHERE   ")
+                .append("       CONVERT (VARCHAR(12), t.cprs, 112) =  ")
                 .append(fillTradeData);
         try {
             List<HashMap<String,String>> resultList = JdbcHelper.query(insertTradeBuilder.toString());
             for (HashMap tradeMap:resultList) {
                 Trade trade = new Trade(tradeMap);
+                Double gross = Double.parseDouble(trade.getGross())/DblinkConnecter.load_ratio;
+                trade.setGross(gross.toString());
+
+                Double tare = Double.parseDouble(trade.getTare())/DblinkConnecter.load_ratio;
+                trade.setTare(tare.toString());
+
+                Double net = Double.parseDouble(trade.getNet())/DblinkConnecter.load_ratio;
+                trade.setNet(net.toString());
                 tradeList.add(trade);
             }
         } catch (SQLException e) {
@@ -116,52 +135,28 @@ public class TradeDataService extends BaseService<Trade, Long> {
             loadometerCodeList.add(loadometerIndicatorDto.getIndicatorCode());
         }
 
-        //删除已经存在的当日的地磅指标数据
-        Searchable removeLoadometerCodeSearchable = Searchable.newSearchable()
-                .addSearchFilter("indicatorCode", MatchType.IN, loadometerCodeList)
-                .addSearchFilter("statDateNum", MatchType.EQ, generateDate);
-        quotaInstanceRepository.deleteInBatch(quotaInstanceRepository.findAll(removeLoadometerCodeSearchable));
+        if (!loadometerCodeList.isEmpty()) {
+            //删除已经存在的当日的地磅指标数据
+            Searchable removeLoadometerCodeSearchable = Searchable.newSearchable()
+                    .addSearchFilter("indicatorCode", MatchType.IN, loadometerCodeList)
+                    .addSearchFilter("statDateNum", MatchType.EQ, generateDate);
+            quotaInstanceRepository.deleteInBatch(quotaInstanceRepository.findAll(removeLoadometerCodeSearchable));
+        }
 
-        log.error("----生成地磅原始指标开始----");
         //保存查询到的当日地磅指标数据
         for(LoadometerIndicatorDto loadometerIndicatorDto:loadometerIndicatorDtoList) {
-            QuotaTemp quotaTemp = quotaTempRepository.findByIndicatorCode(loadometerIndicatorDto.getIndicatorCode());
-            QuotaInstance quotaInstance = new QuotaInstance(quotaTemp);
+            QuotaTemp indicatorTemp = quotaTempRepository.findByIndicatorCode(loadometerIndicatorDto.getIndicatorCode());
+            QuotaInstance quotaInstance = new QuotaInstance(indicatorTemp);
             try {
                 quotaInstance.setFloatValue(Double.parseDouble(loadometerIndicatorDto.getTotalAmount()));
                 quotaInstance.setValueType(QuotaConsts.VALUE_TYPE_DOUBLE);
                 quotaInstance.setStatDateNum(Integer.parseInt(generateDate));
                 quotaInstance.setInstanceTime(DateUtil.stringToDate(generateDate, DateUtil.DATE_FORMAT_DAFAULT));
-                quotaInstance.setCreateTime(Calendar.getInstance());
             } catch (ParseException e) {
                 log.error("stringToCalendar error:", e);
             }
             quotaInstanceList.add(quotaInstance);
         }
-        log.error("----生成地磅原始指标结束----");
         quotaInstanceRepository.save(quotaInstanceList);
-
-        log.error("----生成地磅扩展指标开始----");
-        computExtendLoadoQuota(loadometerCodeList, generateDate);
-        log.error("----生成地磅扩展指标结束----");
-    }
-
-    /**
-     * 计算地磅的扩展指标
-     * @param loadometerCodeList 原始地磅指标
-     * @param generateDate 生成日期
-     */
-    public void computExtendLoadoQuota(List<String> loadometerCodeList, String generateDate) {
-        Calendar computCal = null;
-        try {
-            computCal = DateUtil.stringToCalendar(generateDate, DateUtil.DATE_FORMAT_DAFAULT);
-        } catch (ParseException e) {
-            e.printStackTrace();
-        }
-
-        Searchable searchable = Searchable.newSearchable()
-                .addSearchFilter("indicatorCode", MatchType.IN, loadometerCodeList);
-        List<QuotaTemp> quotaTempList = quotaTempRepository.findAll(searchable).getContent();
-        quotaComputInsService.reComputQuota(computCal,quotaTempList);
     }
 }
